+++ R/ggsegJHU-package.R
@@ -1,0 +1,7 @@
+#' @keywords internal
+#' @import ggseg.formats
+"_PACKAGE"
+
+globalVariables(c(
+  ".jhu_tracts", ".jhu_labels", ".icbm"
+))